pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-pentest-platform"
version = "1.0.0"
authors = [
    { name="Mike", email="mike@example.com" },
]
description = "AI-driven penetration testing platform"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "PyQt5>=5.15.0",
    "python-nmap>=0.7.1",
    "reportlab>=4.0.0",
    "SQLAlchemy>=2.0.0",
    "requests>=2.25.0",
]

[tool.setuptools]
packages = ["core", "gui"] 